+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Added ingredient to formula backend update
 - Fix invalid formula update date on empty formulas
 - Update empty table message
+- Rename sex to gender
+- Error handling improvements
 
 ### Version 11.9
 - Added system logs access via the UI for docker/cloud installations - this comes disabled by default